Fix appendRcsKappa_0 to work with systs
This also introduces a way of skipping values of kappa that are 0 and
instead takes the value from 2017. This expects of course that the
values for 2017 are already calculated and the directories with the root
files have the same pattern except for the year number.
Since the 0b SB is not identical for all the years, the searchbin file
was renamed.