Use W+Rare bkg for Rcs SB calculation, instead of EWK
Since it makes no sense to add all the MC and then subtract the TTJets
afterwards using a complicated formula, instead just add WJets and all
other bkg instead of EWK and use the simple N_SR/N_CR formula in the
WJets SB.
The QCD estimation now automatically reads the correct f-ratio file.
The program now requires the year number (2016, 2017 or 2018) to be part
of the pattern or it exists before execution.
This way one can be sure that things work correctly all the time and one
does not have to manually adjust hardcoded input files or replicate the
files for each year.

The change in the searchBins_0b_with_NW.py is just commenting out a few
lines that can be helpful when using the findMissingBins.sh script.
Since they require manual intervention to use anyway, it makes sense to
have them as a comment per default. It is not a meaningful change, just
more conveninet by default.