Update the sample links to V3
This change also contains a change to only produce the missing bins.
This is especially useful for producing the systematics variations.
The idea is to use findMissingBins to check which bins are missing and
write that to a file. If such a file exists, it will only produce the
bins listed in that file. For producing everything, make sure to delete
the file produced by findMissingBins.