Fix the signal systematics
The readyields did not properly read out the histograms with "syst" in
their name. Now it add the relative uncertainties in quadrature.
Calcsyst was sped up significantly by just using the functions of the
histogram, instead of doing unnecessary slow loops in python.
The tables for the signal are now written in a simpler way.